requirements.txt
attrs
fastapi
uvicorn
pyjwt
python-multipart
requests
celery
passlib
tenacity
pydantic
emails
raven
gunicorn
psycopg2-binary
alembic
sqlalchemy
bcrypt
injector
